requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
fsspec==2025.3.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
uvicorn==0.34.0
